fix: upgrade process

+++ bin/restore-from-restore-folder.sh
@@ -13,13 +13,18 @@ then
 fi
 
 echo "RESTORING DATABASE WITH DATA FROM ${PG_BACKUP_DUMP_RESTORE}"
+rm -rf "${PGDATA}"
+rm -rf "${PG_BACKUP_BASEBACKUP}"/*
+rm -rf "${PG_BACKUP_WAL}"/*
+
+/usr/local/bin/init-db.sh &> /dev/null
+
 runuser --user postgres -- "${PG_BIN}/pg_ctl" -D "${PGDATA}" -o "-c listen_addresses='' -p '5432'"  --wait --timeout="${DATABASE_CHECK_TIME}" --silent --log=/dev/null start
 
-# start with a clean database
-runuser --user postgres -- /usr/bin/dropdb --if-exists --username="${POSTGRES_USER}" --no-password "${POSTGRES_DB}"
-runuser --user postgres -- /usr/bin/createdb --owner="${POSTGRES_USER}" --user="${POSTGRES_USER}" --no-password "${POSTGRES_DB}"
-
-runuser --user postgres -- find "${PG_BACKUP_DUMP_RESTORE}" -type f -name "*.tar" -exec "${PG_BIN}/pg_restore" --create --clean --if-exists --dbname="${POSTGRES_DB}" --username="${POSTGRES_USER}" --no-password "{}" \;
-rm -rf "${PG_BACKUP_DUMP_RESTORE}"/*
+runuser --user postgres -- find "${PG_BACKUP_DUMP_RESTORE}" -type f -name "*.tar" -exec "${PG_BIN}/pg_restore" --dbname="${POSTGRES_DB}" --username="${POSTGRES_USER}" --no-password "{}" \;
+if [ "$?" = 0 ] 
+then
+  rm -rf "${PG_BACKUP_DUMP_RESTORE}"/*
+fi
 
 runuser --user postgres -- "${PG_BIN}/pg_ctl" -D "${PGDATA}" -m fast   --wait --timeout="${DATABASE_CHECK_TIME}" --silent stop
